# order-service/auth-service/api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce API Assignment",
    "version": "1.0",
    "description": "API Documentation for E-Commerce App — Auth & Security Service"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Environment"
    }
  ],
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT via Authorization header (Bearer token)"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "JWT via access_token cookie"
      }
    },
    "schemas": {
      "GenderEnum": {
        "type": "string",
        "enum": ["MALE", "FEMALE"]
      },
      "StatusEnum": {
        "type": "string",
        "enum": ["ACTIVE", "INACTIVE", "DELETED"]
      },
...
      "AdminCreateAccountRequestDTO": {
        "type": "object",
        "properties": {
          "email":       { "type": "string", "format": "email" },
          "phoneNumber": { "type": "string" },
          "password":    { "type": "string" },
          "firstName":   { "type": "string" },
          "lastName":    { "type": "string" },
          "gender":      { "$ref": "#/components/schemas/GenderEnum" },
          "birthDate":   { "type": "string", "format": "date" },
          "roleIds": {
            "type": "array",
            "items": { "type": "string", "format": "uuid" }
          }
        }
      },
...
      "PermissionResponseDTO": {
        "type": "object",
        "properties": {
          "id":     { "type": "string", "format": "uuid" },
          "module": { "type": "string" },
          "code":   { "type": "string" }
        }
      },
      "RoleResponseDTO": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "code":        { "type": "string" },
          "name":        { "type": "string" },
          "description": { "type": "string" },
          "permissions": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PermissionResponseDTO" }
          }
        }
      },
      "AccountResponseForAdminDTO": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "email":       { "type": "string" },
          "phoneNumber": { "type": "string" },
          "status":      { "$ref": "#/components/schemas/StatusEnum" },
          "createdAt":   { "type": "string", "format": "date-time" },
          "updatedAt":   { "type": "string", "format": "date-time" },
          "firstName":   { "type": "string" },
          "lastName":    { "type": "string" },
          "gender":      { "$ref": "#/components/schemas/GenderEnum" },
          "birthDate":   { "type": "string", "format": "date" },
          "roles": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RoleResponseDTO" }
          }
        }
      },
...
      "ApiResponseAccountAdmin": {
        "type": "object",
        "properties": {
          "code":      { "type": "integer" },
          "message":   { "type": "string" },
          "result":    { "$ref": "#/components/schemas/AccountResponseForAdminDTO" },
          "timestamp": { "type": "string", "format": "date-time" },
          "path":      { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/accounts/admin/{id}": {
      "put": {
        "tags": ["Accounts"],
        "summary": "Update account (Admin)",
        "description": "Requires role `ADMIN` or permission `account:write`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AdminCreateAccountRequestDTO" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAccountAdmin" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Accounts"],
        "summary": "Get account by ID (Admin)",
        "description": "Requires role `ADMIN` or permission `account:read`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Account fetched",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponseAccountAdmin" }
              }
            }
          }
        }
      }
    },
...
  }
}